DRY up the view_request and view_batch_request integration specs
